requirements.txt
flask
flask-cors
gunicorn
discord
python-dotenv
flask-sqlalchemy
psycopg2
requests
aiofiles